requirements.txt
python-dotenv
tenacity
openai